+++ lib/flutter_metrics_sdk.dart
@@ -1,6 +1,8 @@
-library flutter_metrics_sdk;
+library;
 
 export 'src/metrics_client.dart';
 export 'src/metrics_event.dart';
 export 'src/screen_tracker.dart';
 export 'src/crash_tracker.dart';
+export 'src/api_tracker.dart';
+export 'src/frame_tracker.dart';